Add CsvTrainerJob.
