+++ brushfire-features/src/main/scala/com/stripe/brushfire/features/package.scala
@@ -1,0 +1,6 @@
+package com.stripe.brushfire
+
+package object features {
+  type FeatureValue = Dispatched[Double, String, Double, String]
+  type CsvRow = IndexedSeq[String]
+}
