+++ build.sbt
@@ -27,14 +27,14 @@ lazy val brushfireCore = project.
   in(file("brushfire-core")).
   disablePlugins(sbtassembly.AssemblyPlugin)
 
+lazy val brushfireFeatures = project.
+  in(file("brushfire-features")).
+  dependsOn(brushfireCore)
+
 lazy val brushfireScalding = project.
   in(file("brushfire-scalding")).
-  dependsOn(brushfireCore)
+  dependsOn(brushfireCore, brushfireFeatures)
 
 lazy val brushfireFinatra = project.
   in(file("brushfire-finatra")).
-  dependsOn(brushfireCore)
-
-lazy val brushfireFeatures = project.
-  in(file("brushfire-features")).
   dependsOn(brushfireCore)
